refactor(consensus): genericize the raft transport, admin, and log adapters
Second step of the fluree-raft-core extraction. The four openraft-facing
adapters move to the core crate, parameterized by the type config
instead of monomorphized against the nameservice's.

New: `FlureeRaftConfig`, the constrained openraft profile. It pins the
five associated types every Fluree group shares — NodeId=u64,
Node=ClusterNode, Entry=Entry<C>, SnapshotData=Cursor<Vec<u8>>, and
Responder=OneshotResponder<C> — so each adapter carries one bound
instead of five. Applications still write their own
`declare_raft_types!`; a blanket impl covers whatever it produces.

The Responder pin is not cosmetic: openraft gates its blocking
client-write API on that exact type, so without it `add_learner` and
`change_membership` do not exist on `Raft<C>` and the whole admin
surface fails to resolve. `declare_raft_types!` already defaults to it.

Moved and genericized:
  - log_adapter — LogAdapter<C, S>. Log-id and vote converters stay
    concrete since NodeId is pinned.
  - network — RaftTransportConfig + client + relative RPC router.
  - admin — RaftAdmin<C> + relative admin router.
  - forward — now generic over a `LeaderView` source rather than
    `Raft` directly. This is what the routing decision actually needs
    (current leader + membership addresses), and it turns seven cases
    that previously required a live 5-node cluster into unit tests:
    election-in-progress, self-is-leader, leader-missing-from-
    membership, and the SSRF guard's loopback and link-local arms.

NetworkConfig splits along the same line it was already fused across.
The generic transport keeps timeouts and the body caps for openraft's
three routes plus leader forwarding; cross_node_propose_timeout and the
apply_staged_commit / apply_queue_poison caps stay in consensus, since
they configure RPCs a crate that knows nothing about staged commits has
no business describing. Consensus's NetworkConfig now embeds the
transport half; the ~6 call sites are updated rather than papered over
with a Deref. Its Default is hand-written because a derived one would
silently zero three real tuning values.

`build_client` becomes a free function: it never used the type config,
and one client is meant to be shared across every group in a process.

Routers stay relative — no prefix of their own — so the host nests them
at the legacy /raft and /cluster or under a group id. The two SSRF-guard
helpers become public because the nameservice's own cross-node forward
validates the same membership-sourced URLs, and a second drifting copy
of that check is the thing to avoid.

openraft is gated behind a new `raft` feature on the core crate, wired
from the consensus `raft` feature. Verified openraft is still absent
from the default build, so monolithic users — who reach this crate only
for http::is_hop_by_hop via peer mode — are unaffected.

423 core+consensus tests, the 14-test server raft suite including the
5-node HTTP cluster (failover, liveness demotion, concurrent writes),
workspace clippy and fmt all clean.
