+++ docs/design/raft-command-queue.md
@@ -21,10 +21,10 @@ The Raft consensus crate (`fluree-db-consensus/src/raft/`) is structured as a se
 | `Command`, `Response`      | `state_machine.rs`                                       | The log entry types. ~20 variants spanning transaction flow, ledger lifecycle, and metadata.                |
 | `NameServiceState`         | `state_machine.rs`                                       | The replicated in-memory state: branch heads, ledger registry, per-branch queues, idempotency cache.        |
 | `StateMachineAdapter`      | `state_machine_adapter.rs`                               | openraft's `RaftStateMachine` impl. Applies entries, takes/installs snapshots, resolves waiters.            |
-| `LogStore`, `SnapshotStore`| `log_adapter.rs`; `fluree-raft-core` `storage/{fs,memory}.rs` | openraft's `RaftLogStorage`/`RaftLogReader`/`RaftSnapshotBuilder`. Local-disk persistence. The backends are openraft-free — they store opaque `Vec<u8>` payloads. |
-| `HttpRaftNetworkFactory`   | `network.rs`                                             | Inter-node RPC (`/raft/vote`, `/raft/append-entries`, `/raft/install-snapshot`) over HTTP.                  |
-| `RaftAdmin` / `/cluster/*` | `admin.rs`                                               | Operator-facing membership endpoints (`initialize`, `add-learner`, `change-membership`, `status`).         |
-| Follower-forward middleware| `forward.rs`                                             | Axum middleware that proxies leader-only client requests to the current leader.                            |
+| `LogStore`, `SnapshotStore`| `fluree-raft-core` `log_adapter.rs`, `storage/{fs,memory}.rs` | openraft's `RaftLogStorage`/`RaftLogReader`/`RaftSnapshotBuilder`. Local-disk persistence. The backends are openraft-free — they store opaque `Vec<u8>` payloads. |
+| `HttpRaftNetworkFactory`   | `fluree-raft-core` `network.rs`                          | Inter-node RPC (`/raft/vote`, `/raft/append-entries`, `/raft/install-snapshot`) over HTTP. Generic over the type config; `network.rs` here keeps only the nameservice-specific `NetworkConfig` fields. |
+| `RaftAdmin` / `/cluster/*` | `fluree-raft-core` `admin.rs`                            | Operator-facing membership endpoints (`initialize`, `add-learner`, `change-membership`, `status`).         |
+| Follower-forward middleware| `fluree-raft-core` `forward.rs`                          | Axum middleware that proxies leader-only client requests to the current leader. Generic over a `LeaderView` source. |
 | `QueuedTransactor`         | `queued_transactor.rs`                                   | Client-side proposer. Builds envelopes, writes to CAS, proposes `EnqueueCommand`, awaits the typed receipt. |
 | `commit_worker::Worker`             | `commit_worker.rs`                                       | Leader-only. Drains per-branch queues, stages work, proposes `ApplyHead`.                                  |
 | `EvictionScheduler`        | `eviction_scheduler.rs`                                  | Leader-only. Periodically proposes `EvictIdempotency` to age out the cache.                                |
